Crate rustpython

Modules

shell

Functions

_run_string

create_settings
Create settings by examining command line arguments and environment variables.

get_env_var_value
Get environment variable and turn it into integer.

get_paths
Helper function to retrieve a sequence of paths from an environment variable.

main

parse_arguments

run_command

run_module

run_rustpython

run_script